remove not working script + some improvement with script for ROUGE's configuration generator

+++ gen_rouge_in.php
@@ -9,10 +9,25 @@
  */
 
 $ROUGE_VERSION = "1.5.5";
+
+if(count($argv) < 3) {
+	echo "USAGE :\n";
+	echo "php gen_rouge_in.php </path/to/models/dir> </path/to/peers/dir> [<peers_extensions>]\n\n";
+	echo "exp1: php gen_rouge_in.php models peers txt >> config.xml\n";
+	echo "exp2: php gen_rouge_in.php models peers ter1,ter2 >> config.xml\n";
+	exit(1);
+}
+
 $MODEL_ROOT = realpath($argv[1]);
 $PEER_ROOT = realpath($argv[2]);
 if(isset($argv[3])) {
-	$PEER_EXTENSIONS = explode(',',$argv[3]);
+	$PEER_EXTENSIONS = array();
+	$tmp_peers_ext = explode(',',$argv[3]);
+	foreach($tmp_peers_ext as $ext) {
+		if($ext !== '') {
+			$PEER_EXTENSIONS[] = $ext;
+		}
+	}
 } else {
 	$PEER_EXTENSIONS = array('ter');
 }
@@ -26,25 +41,36 @@ $models = sort_models(scandir($MODEL_ROOT));
 echo "<ROUGE_EVAL version=\"$ROUGE_VERSION\">\n";
 
 foreach($models as $m_class => $ms) {
-	echo "<EVAL ID=\"$m_class\">\n";
-	echo "<PEER-ROOT>$PEER_ROOT</PEER-ROOT>\n";
-	echo "<MODEL-ROOT>$MODEL_ROOT</MODEL-ROOT>\n";
-	echo "<INPUT-FORMAT TYPE=\"$INPUT_FORMAT_TYPE\"></INPUT-FORMAT>\n";
-	echo "<PEERS>\n";
+	$find = false;
 	foreach($PEER_EXTENSIONS as $p) {
-		echo "<P ID=\"$p\">$m_class.$p</P>\n";
+		if(file_exists($PEER_ROOT.'/'.$m_class.'.'.$p)) {
+			$find = true;
+			break;
+		}
 	}
-	echo "</PEERS>\n";
-	echo "<MODELS>\n";
-	foreach($ms as $m_id => $m) {
-		echo "<M ID=\"$m_id\">$m</M>\n";
+	if($find) {
+		echo "<EVAL ID=\"$m_class\">\n";
+		echo "<PEER-ROOT>$PEER_ROOT</PEER-ROOT>\n";
+		echo "<MODEL-ROOT>$MODEL_ROOT</MODEL-ROOT>\n";
+		echo "<INPUT-FORMAT TYPE=\"$INPUT_FORMAT_TYPE\"></INPUT-FORMAT>\n";
+		echo "<PEERS>\n";
+		foreach($PEER_EXTENSIONS as $p) {
+			if(file_exists($PEER_ROOT.'/'.$m_class.'.'.$p)) {
+				echo "<P ID=\"$p\">$m_class.$p</P>\n";
+			}
+		}
+		echo "</PEERS>\n";
+		echo "<MODELS>\n";
+		foreach($ms as $m_id => $m) {
+			echo "<M ID=\"$m_id\">$m</M>\n";
+		}
+		echo "</MODELS>\n";
+		echo "</EVAL>\n";
 	}
-	echo "</MODELS>\n";
-	echo "</EVAL>\n";
 }
 
 echo "</ROUGE_EVAL>\n";
-
+// /home/anthony/Dropbox/Cours/TER/ter-2015/peers/D0848-A.M.100.H.txt.1
 /* *** FUNCTIONS *** */
 
 function sort_models($in_models) {
